\name{SumVg-package}
\alias{SumVg-package}
\alias{SumVg}
\docType{package}
\title{
Estimate the total heritability explained by all variants in GWAS, with estimation of standard error
}
\description{
The program estiamtes the total heritability explained by all variants in GWAS. Standard errors were computed by resampling methods.
}
\details{
\tabular{ll}{
Package: \tab SumVg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2023-05-31\cr
License: \tab CC BY-NC-SA 4.0\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Hon-Cheong So
}
\references{
So, H.C., Li, M. and Sham, P.C. (2011) Uncovering the total heritability explained by all true susceptibility variants in a genome-wide association study
Genet Epidemiol, 35, 447-456.
\cr So, H.C., Xue, X. and Sham, P.C. (2023) SumVg: Total heritability explained by all variants in genome-wide association studies based on summary statistics with standard error estimates.
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
